requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
async-timeout=4.0.1=pypi_0
atomicwrites=1.4.0=py_0
attrs=21.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bottleneck=1.3.2=py39h7cc1a96_1
ca-certificates=2021.10.26=haa95532_2
cachetools=4.2.4=pypi_0
certifi=2021.10.8=py39haa95532_2
charset-normalizer=2.0.9=pypi_0
click=8.0.3=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
configparser=5.2.0=pypi_0
cudatoolkit=10.2.89=h74a9793_1
docker-pycreds=0.4.0=pypi_0
freetype=2.10.4=hd328e21_0
frozenlist=1.2.0=pypi_0
fsspec=2021.11.1=pypi_0
future=0.18.2=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.26=pypi_0
google-auth=2.3.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.42.0=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.8.2=pypi_0
iniconfig=1.1.1=pyhd3eb1b0_0
intel-openmp=2021.4.0=haa95532_3556
jpeg=9d=h2bbff1b_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.0=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
markdown=3.3.6=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
more-itertools=8.12.0=pyhd3eb1b0_0
multidict=5.2.0=pypi_0
numexpr=2.8.1=py39hb80d3ca_0
numpy=1.21.2=py39hfca59bb_0
numpy-base=1.21.2=py39h0829f74_0
oauthlib=3.1.1=pypi_0
olefile=0.46=pyhd3eb1b0_0
openssl=1.1.1m=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py39h6214cd6_0
pathtools=0.1.2=pypi_0
pillow=8.4.0=py39hd45dc43_0
pip=21.2.4=py39haa95532_0
pluggy=0.13.1=py39haa95532_0
promise=2.3=pypi_0
protobuf=3.19.1=pypi_0
psutil=5.9.0=pypi_0
py=1.10.0=pyhd3eb1b0_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pydeprecate=0.3.1=pypi_0
pyparsing=3.0.6=pypi_0
pytest=6.2.4=py39haa95532_2
python=3.9.7=h6244533_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.10.0=py3.9_cuda10.2_cudnn7_0
pytorch-lightning=1.5.5=pypi_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=pypi_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.8=pypi_0
sentry-sdk=1.5.2=pypi_0
setuptools=58.0.4=py39haa95532_0
shortuuid=1.0.8=pypi_0
six=1.16.0=pyhd3eb1b0_0
smmap=5.0.0=pypi_0
sqlite=3.36.0=h2bbff1b_0
subprocess32=3.5.4=pypi_0
tensorboard=2.7.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.11=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
torchmetrics=0.6.1=pypi_0
torchtyping=0.1.4=pypi_0
torchvision=0.11.1=py39_cu102
tqdm=4.62.3=pypi_0
typeguard=2.13.3=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
urllib3=1.26.7=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wandb=0.12.9=pypi_0
werkzeug=2.0.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
wincertstore=0.2=py39haa95532_2
xz=5.2.5=h62dcd97_0
yarl=1.7.2=pypi_0
yaspin=2.1.0=pypi_0
zipp=3.6.0=pypi_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
